update readme and move legacy code to legacy folder
