+++ frontend/src/lib/tools/legacy/semantic-documentation-search.ts
