+++ frontend/src/lib/tools/legacy/velocity-management.ts
